重构Scan函数

+++ run/fetcher.py
@@ -3,15 +3,15 @@ import configparser as cp
 
 import sys
 import os
+
 sys.path.append(os.path.dirname(sys.path[0]))
 
 import getopt
-import datetime
 import utils.db_tool
 import utils.misc
+import datetime
 
 from jqdatasdk import *
-
 
 JK_User = None
 JK_Token = None
@@ -44,21 +44,44 @@ def _scan():
         print("Last Round NOT Finished,Exit")
         return
     # 只能在Linux上运行
-    os.mknod(File_Locked)
+    # os.mknod(File_Locked)
     # 开始Scan，目前只有Daily抓取
     fp = open(File_TBF, 'w')
-    tds = set(Stock_DB_Tool.get_trade_days('2013-01-01', str(datetime.date.today())))
-    sids = Stock_DB_Tool.get_stock_info(['stock_id'])
-    for sid in sids:
-        suffix = utils.misc.stockid2table(sid[0], 10)
+    stocks = Stock_DB_Tool.get_stock_info(['stock_id', 'start_date'])
+    for stock in stocks:
+        print(stock)
+        stock_id = stock[0]
+        ipo_date = stock[1]
+        start_date = datetime.datetime.strptime('2013-01-01', '%Y-%m-%d').date()
+        if ipo_date > start_date:
+            start_date = ipo_date
+        tds = set(Stock_DB_Tool.get_trade_days(str(start_date), str(datetime.date.today())))
+        suffix = utils.misc.stockid2table(stock_id, 10)
         table_name = "quant_stock.price_daily_r" + str(suffix)
-        sql = "select dt from " + table_name + " where sid = \'" + sid[0] + '\''
+        sql = "select dt from " + table_name + " where sid = \'" + stock_id + '\''
         dt = set(Stock_DB_Tool.exec_raw_select(sql))
         tbf_fetch_list = tds - dt
         for tbf in tbf_fetch_list:
-            print(sid[0] + ',' + str(tbf[0]) + ',Daily', file=fp)
-        print(sid[0] + ' Finished.')
+            print(stock_id + ',' + str(tbf[0]) + ',Daily', file=fp)
+        print(stock_id + " Finished.")
     fp.close()
+    # 释放锁
+    if os.path.exists(File_Locked):
+        os.remove(File_Locked)
+
+
+def _fetch_price():
+    if not os.path.exists(TBF_Dir):
+        print("TBF Dir NOT Exit")
+        return
+    os.chdir(TBF_Dir)
+    # 确认锁及抓取文件
+    if os.path.exists(File_Locked) or not os.path.exists(File_TBF):
+        print("Last Round NOT Finished OR NO TBF,Exit")
+        return
+    # 只能在Linux上运行
+    os.mknod(File_Locked)
+
     # 释放锁
     if os.path.exists(File_Locked):
         os.remove(File_Locked)
@@ -100,10 +123,10 @@ if __name__ == '__main__':
                 _check_spare()
             elif opt == '--scan':
                 # 获取待抓取的列表
-                print("Scan For Price")
                 _scan()
             elif opt == '--price':
                 # 抓取行情
+                _fetch_price()
                 print("Fetch Price")
             else:
                 raise getopt.GetoptError("Usage Error")
